Integrate LazyMapList into Lists2
Should help discover the functionality, as it's closer to the other
`Lists2.map` related functions we have.
